ml_backend/requirements.txt
Django==4.1.7
djangorestframework==3.14.0
tensorflow==2.12.0
tensorflow-io==0.32.0
scipy==1.9.0
scikit-image==0.20.0
django-cors-headers==4.1.0